+++ drivers/gpu/drm/mediatek/mediatek_v2/mtk_drm_ddp_comp.c
@@ -1179,7 +1179,7 @@ void mtk_ddp_comp_clk_prepare(struct mtk_ddp_comp *comp)
 
 	if (comp->mtk_crtc)
 		index = drm_crtc_index(&comp->mtk_crtc->base);
-	CRTC_MMP_MARK(index, ddp_clk, comp->id, 1);
+	CRTC_MMP_MARK((int) index, ddp_clk, comp->id, 1);
 }
 
 void mtk_ddp_comp_clk_unprepare(struct mtk_ddp_comp *comp)
